Changed pixels from i32 to f32

+++ src/zui/render_layer.rs
@@ -8,14 +8,14 @@ pub struct RenderLayer {
     pub text_vertices: Vec<TextVertex>,
 
     /// The clipping rectangle, contents outside of this will not be rendered
-    pub clip_rectangle: Option<Rectangle<i32>>,
+    pub clip_rectangle: Option<Rectangle<f32>>,
 }
 
 impl RenderLayer {
     pub fn new(
         simple_vertices: Vec<SimpleVertex>,
         text_vertices: Vec<TextVertex>,
-        clip_rectangle: Option<Rectangle<i32>>,
+        clip_rectangle: Option<Rectangle<f32>>,
     ) -> Self {
         Self {
             simple_vertices,
